add errors handle

+++ src/entities/category.js
@@ -1,10 +1,15 @@
 const Category = require("../models/category");
+const ErrorService = require("../middleware/error/errorService");
 
 async function getAllCategories() {
   return Category.query().select();
 }
 
 async function getCategoryById(id) {
+  const category = await Category.query().findById(id);
+  if (!category) {
+    throw ErrorService.errorThrow(404);
+  }
   return Category.query().findById(id);
 }
 
@@ -13,10 +18,18 @@ async function createCategory(post) {
 }
 
 async function updateCategory(id, data) {
+  const category = await Category.query().findById(id);
+  if (!category) {
+    throw ErrorService.errorThrow(404);
+  }
   return Category.query().findById(id).patch(data);
 }
 
 async function removeCategoryById(id) {
+  const category = await Category.query().findById(id);
+  if (!category) {
+    throw ErrorService.errorThrow(404);
+  }
   return Category.query().findById(id).delete();
 }
